Criação das pastas de referências e inicialização estrutura
