+++ .vscode/launch.json
@@ -1,0 +1,30 @@
+{
+    // Use o IntelliSense para saber mais sobre os atributos possíveis.
+    // Focalizar para exibir as descrições dos atributos existentes.
+    // Para obter mais informações, acesse: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        
+        
+        {
+            "name": "app_movies",
+            "cwd": "app_movies",
+            "request": "launch",
+            "type": "dart"
+        },
+        {
+            "name": "app_movies (profile mode)",
+            "cwd": "app_movies",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "profile"
+        },
+        {
+            "name": "app_movies (release mode)",
+            "cwd": "app_movies",
+            "request": "launch",
+            "type": "dart",
+            "flutterMode": "release"
+        }
+    ]
+}
