+++ app_movies/lib/core/app_colors.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class AppColors{
+  static const Color primaryColor = Color(0xFF024873);
+  static const Color secondaryColor = Color(0xFF0381A9);
+  static const Color tertiaryColor = Color(0xFF2FB7AD);
+  static const Color quaternaryColor = Color(0xFFAECDE2);
+
+  // static const Color backgroundColor = Color ();
+  // static const Color accentColor = Color();
+  // static const Color textColor = Color();
+
+}
