+++ app_movies/lib/core/app_images.dart
@@ -1,0 +1,3 @@
+class AppImages {
+  static const String teste = 'assets/images/teste.png';
+}
