+++ app_movies/lib/core/app_widget.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:app_movies/pages/Home/home_page.dart';
+
+
+
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      initialRoute: '/home',
+      routes: {
+        '/home': (context) => const MyHomePage(),
+      },
+    );
+  }
+}
