+++ app_movies/lib/pages/Home/home_page.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:app_movies/core/app_images.dart';
+
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Aplicativo de Filmes'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'Bem-vindo ao Aplicativo de Filmes',
+              style: TextStyle(fontSize: 24),
+            ),
+            SizedBox(height: 20),
+            Image.asset(
+              AppImages.teste,
+              width: 150,
+              height: 150,
+              fit: BoxFit.contain,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
